+++ src/main/java/com/verdantartifice/thaumicwonders/common/containers/slots/SlotPrimordialPearl.java
@@ -1,13 +1,13 @@
 package com.verdantartifice.thaumicwonders.common.containers.slots;
 
 import com.verdantartifice.thaumicwonders.common.items.ItemsTW;
-import net.minecraft.inventory.IInventory;
-import net.minecraft.inventory.Slot;
 import net.minecraft.item.ItemStack;
+import net.minecraftforge.items.IItemHandler;
+import net.minecraftforge.items.SlotItemHandler;
 
-public class SlotPrimordialPearl extends Slot {
-    public SlotPrimordialPearl(IInventory tileEntity, int index, int xPosition, int yPosition) {
-        super(tileEntity, index, xPosition, yPosition);
+public class SlotPrimordialPearl extends SlotItemHandler {
+    public SlotPrimordialPearl(IItemHandler stackHandler, int index, int xPosition, int yPosition) {
+        super(stackHandler, index, xPosition, yPosition);
     }
 
     @Override